updated references, fixed issues with data manager user and schema versions

+++ trogz.game.data.lib/datamanager.h
@@ -16,8 +16,8 @@
     NSBundle* bundle;
     NSString* resourcePath;
     
-    int resourceUserVersion;
-    int resourceSchemaVersion;
+    long resourceUserVersion;
+    long resourceSchemaVersion;
 }
 
 @property (nonatomic, assign) bool needsDataRefresh;
